atualizado

+++ src/app/setor/setor.component.ts
@@ -1,7 +1,9 @@
 import {Component, OnInit} from '@angular/core';
 import {SetorService} from './setor.service';
 import {Setor} from './setor';
-import {LoginService} from '../login/login.service'; 
+import {Message} from 'primeng/api';
+import {LoginService} from '../login/login.service';
+import {ConfirmationService} from 'primeng/api';
 
 @Component({
   templateUrl: './setor.component.html',
@@ -11,15 +13,17 @@ export class SetorComponent implements OnInit {
 
   setores: Setor[];
   showDialog = false;
+  showConfirm = false;
   setorEdit = new Setor();
+  msgs: Message[] = [];
 
-  constructor(private setorService: SetorService, private loginService: LoginService) {
-  }
+  constructor(private setorService: SetorService, private loginService: LoginService, 
+  private confirmationService: ConfirmationService) {}
+
 
   hasRole(role: string): boolean {
     return this.loginService.hasRole(role);
   }
-
 
   ngOnInit(): void {
     this.findAll();
@@ -39,17 +43,38 @@ export class SetorComponent implements OnInit {
       this.setorEdit = new Setor();
       this.findAll();
       this.showDialog = false;
+      this.msgs = [{severity:'sucess', summary:'Confirmado', detail:'Registro salvo com sucesso'}];      
+    },
+    error => {
+      this.msgs = [{severity:'error', summary:'Erro', detail:'Certifique-se de preencher todos os campos.'}];
     });
   }
 
   editar(setor: Setor) {
     this.setorEdit = setor;
     this.showDialog = true;
+    //this.msgs = [{severity:'sucess', summary:'Confirmado', detail:'Registro alterado com sucesso'}];
   }
 
-  remover(setor: Setor) {
-    this.setorService.delete(setor.id).subscribe(() => {
-      this.findAll();
-    });
+  confirmDelete(setor: Setor) {
+        this.confirmationService.confirm({
+            message: 'Essa ação não poderá ser desfeita',
+            header: 'Deseja remover esse registro?',
+           
+            accept: () => {
+                this.setorService.delete(setor.id).subscribe(() => {
+                this.findAll();
+                this.msgs = [{severity:'sucess', summary:'Confirmado', detail:'Registro removido com sucesso'}];
+              },
+              error => {
+                this.msgs = [{severity:'error', summary:'Erro', detail:'Este registro nao pode ser removido.'}];
+              });
+            }
+        });
+    }
+
+  cancelar(){
+    this.showDialog = false;
+    this.setorService.findAll().subscribe(e => this.setores = e);
   }
 }
